fix pr review

+++ Engine/src/Runtime/RHI/Exar/ExarBufferView.cpp
@@ -3,7 +3,7 @@
 
 
 UnasciiEngine::RHI::EXAR::ExarBufferView::ExarBufferView(const ExarBuffer& pBuffer)
-	: mData(reinterpret_cast<const u8*>(pBuffer.getCPUAddress()), static_cast<size_t>(pBuffer.getDesc().size))
+	: mData(reinterpret_cast<const u8*>(pBuffer.getCPUAddress()), pBuffer.getDesc().size)
 	, mCursor(0)
 {
 }
